Lint: first pass over the landed protocol work
Fixes golangci-lint findings in the work landed today: gofmt, US
spelling, comment periods, lower-case error strings, errors.Is and %w
wrapping, checked type assertions in tests, exhaustive switch defaults,
parallel subtests without shared state, tagged switches, removed
conversions, gocritic rewrites, context-bound listeners, preallocation,
and named l2tp scenario constants. buildPathErr loses its unused hop
parameter, since an RFC 2205 PathErr carries no RSVP_HOP. RFC 1195
spells prefered in a verbatim quote, so misspell excludes that one path.
Remaining lint findings and four failing test packages are still open.

+++ internal/component/ike/engine/responder.go
@@ -528,7 +528,7 @@ func (ps *PeerSession) handleAuthRequest(sa *SA, msg *wire.Message, rawMsg []byt
 	// RFC 4555 Section 3.9 protects the first AUTH independently of whether the
 	// peer advertises MOBIKE. An absent capability cannot waive local NAT policy.
 	if sa.PeerCfg.ProhibitNAT && (sa.NATDetected || notifyOf(inner, wire.NotifyNoNATsAllowed) == nil) {
-		notify := uint16(wire.NotifyAuthenticationFailed)
+		notify := wire.NotifyAuthenticationFailed
 		// RFC 7296 Section 2.21.2 forbids fatal extension errors until the
 		// peer has shown it understands them. Either RFC 4555 notification
 		// demonstrates that understanding; policy alone cannot manufacture it.
